+++ lib/network-exporters/formatters/graphml/helpers.ts
@@ -1,6 +1,13 @@
-import { type NcEntity } from '@codaco/shared-consts';
+import { type NcEdge, type NcNode } from '@codaco/shared-consts';
+import { type Document } from '@xmldom/xmldom';
+import { createHash } from 'crypto';
 import { isNil } from 'lodash';
 import { getEntityAttributes } from '../../utils/general';
+
+// Utility sha1 function that returns hashed text
+export const sha1 = (text: string) => {
+  return createHash('sha1').update(text, 'utf8').digest('hex');
+};
 
 /**
  * For a given key, return a valid Graphml data 'type' for encoding
@@ -15,7 +22,7 @@ import { getEntityAttributes } from '../../utils/general';
  * @param {*} data
  * @param {*} key
  */
-const getGraphMLTypeForKey = (data: NcEntity[], key: string) =>
+export const getGraphMLTypeForKey = (data: NcNode[] | NcEdge[], key: string) =>
   data.reduce((result, value) => {
     const attrs = getEntityAttributes(value);
     if (isNil(attrs[key])) return result;
@@ -36,32 +43,18 @@ const getGraphMLTypeForKey = (data: NcEntity[], key: string) =>
     return 'string';
   }, '');
 
-const createElement = (
-  xmlDoc: XMLDocument,
-  tagName: string,
-  attrs: Record<string, string>,
-  child = null,
-) => {
-  const element = xmlDoc.createElement(tagName);
-  Object.entries(attrs).forEach(([key, val]) => {
-    element.setAttribute(key, val);
-  });
-  if (child) {
-    element.appendChild(child);
-  }
-  return element;
-};
-
-const createDataElement = (
-  xmlDoc: XMLDocument,
+export const createDataElement = (
+  xmlDoc: Document,
   attributes: Record<string, string>,
   text: string,
 ) => {
   const textNode = xmlDoc.createTextNode(text);
-  const element = createElement(xmlDoc, 'data', attributes);
+  const element = xmlDoc.createElement('data');
+  Object.entries(attributes).forEach(([key, val]) => {
+    element.setAttribute(key, val);
+  });
+
   element.appendChild(textNode);
 
   return element;
 };
-
-export { createDataElement, getGraphMLTypeForKey };
